transcription fixed

+++ payment-parent/payment-core/src/main/java/com/progressoft/jip/payment/transcription/Transcription.java
@@ -4,6 +4,6 @@ public interface Transcription {
 
 	String getSupportedLanguage();
 
-	String transcript(long number);
+	String transcript(float number);
 
 }
